Colors are stupid, but now tints work
Comfy's GPU tinting works in linear space, and while textures are being converted from sRGB to linear space, the colors passed to shaders are passed as in. Our shaders then do `a*b`, which means we're implicitly doing tinting in linear space.

But blood canvas runs on the CPU and uses image data from the `image` crate, which apparently works in sRGB at least in the way we're using it. This means that the same tint color we get from `to_quad_draw(...)` can't just multiply, as that would be sRGB * linear. For simplicity we'll just convert it to sRGB and then do .linear_space_tint(...), which does a conversion to linear space, multiplication, and conversion back. We could do this differently, but this works for now.